Move pkg/shutdown to internal/shutdown
