+++ cmd/main.go
@@ -13,7 +13,7 @@ import (
 	"github.com/heroticket/internal/infra/mongo"
 	"github.com/heroticket/internal/infra/redis"
 	"github.com/heroticket/internal/jwt"
-	"github.com/heroticket/pkg/shutdown"
+	"github.com/heroticket/internal/shutdown"
 	"go.uber.org/zap"
 
 	_ "github.com/joho/godotenv/autoload"
